chore(ci): clear ruff/mypy so progress PR can merge
Fix lint/type issues blocking CI: ProgressLike protocol, import order,
complexity noqas, and test typing so ruff+mypy+pytest are green.

+++ src/nq/orderbook/depth.py
@@ -113,7 +113,7 @@ def unpack_levels_from_row(
         if sz_i <= 0:
             continue
         # الأعمدة مخزّنة بسعر حقيقي؛ نعيد للنقطة الثابتة للمسح الداخلي
-        levels.append((int(round(px_f / PRICE_SCALE)), sz_i))
+        levels.append((round(px_f / PRICE_SCALE), sz_i))
     return levels
